Add -c to cache symbol information to disk
By far we spend the most time parsing dwarf data. This commit adds a -c
option to cache data to disk or read it from disk, substantially
speeding up the program.

Note that we could still be smarter. If we stored the data in a btree,
for example, the lookups would be faster. But this is good enough for
now.