Use wrappers for _read and _write to check bytes
Ok, we need to stop being lazy and actually check to see if read and
write got all the bytes. This commit changes the read and write calls to
use wrappers that keep reading/writing until all of the bytes are
handled.

This alos means we can change all of the return value checks to just
seeing if they're < 0.