Properly match when address is exactly address of symbol
When the address given was exactly the same as the start of the function
we were returning the wrong function (off-by-one somethingsomething).

This case probably isn't too common in the real world, because crashing
on the very first instruction of a function isn't going to happen often,
but it was still wrong.