Fall back on using symbols if dwarf aranges don't have an address
There are some cases (for whatever reason) where symbols fall outside of
the range of all DWARF aranges. Whether this is an issue with the
generation of DWARF data I'm not sure, but the best thing to do in this
case is likely to fall back on using symbol addresses. Similar to with
the symbol table, if an address is after the start of a symbol but
before the next symbol we print out the name of that first symbol.

Note that in these cases it seems there is disagreement about what
symbol to use depending on what program and what method you use. For
example, old versions of atos don't show any symbol, but newer versions
do. Trying to 'list' one of these addresses doesn't show anything but
using 'image lookup' does. My guess is that some programs are doing a
similar fallback and others aren't.