Use LC_SYMTAB directly instead of LC_FUNCTION_STARTS
We were using a combination of the LC_FUNCTION_STARTS section and the
LC_SYMTAB section to resolve addresses in non-DWARF files. I'm not
really sure we were doing this, I think it was an optimization. Turns
out that not all binaries populate LC_FUNCTION_STARTS. (I saw it in an
MH_DYLIB_STUB file).

Here I'm dropping all LC_FUNCTION_STARTS and just using the symbol
table. I'm not yet sure if this is a good idea, or if this should only
be a fallback.